+++ assets/javascripts/kitten/components/atoms/icon-badge/index.js
@@ -15,7 +15,7 @@ const StyledBadge = styled.span`
   padding: 0;
   min-width: ${pxToRem(30)};
   min-height: ${pxToRem(30)};
-  border-radius: ${pxToRem(30)};
+  border-radius: var(--border-radius-rounded);
   background-color: var(--background-color, ${COLORS.primary1});
 
   &.k-IconBadge--empty {
@@ -34,7 +34,6 @@ const StyledBadge = styled.span`
   &.k-IconBadge--tiny {
     min-width: ${pxToRem(20)};
     min-height: ${pxToRem(20)};
-    border-radius: ${pxToRem(20)};
 
     .k-IconBadge__content {
       font-size: ${stepToRem(-1)};
@@ -49,7 +48,6 @@ const StyledBadge = styled.span`
   &.k-IconBadge--big {
     min-width: ${pxToRem(40)};
     min-height: ${pxToRem(40)};
-    border-radius: ${pxToRem(40)};
 
     .k-IconBadge__content {
       font-size: ${stepToRem(-1)};
@@ -59,7 +57,6 @@ const StyledBadge = styled.span`
   &.k-IconBadge--huge {
     min-width: ${pxToRem(50)};
     min-height: ${pxToRem(50)};
-    border-radius: ${pxToRem(50)};
 
     .k-IconBadge__content {
       font-size: ${stepToRem(-1)};